pyproject.toml
[project]
name = "epilogue-site"
version = "0.1.0"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.1.6",
    "django-anymail[postmark]>=12.0",
    "django-cors-headers>=4.7.0",
    "django-impersonate>=1.9.4",
    "django-mathfilters>=1.0.0",
    "django-ratelimit>=4.1.0",
    "gunicorn>=23.0.0",
    "huey>=2.5.2",
    "markdown>=3.7",
    "psycopg>=3.2.4",
    "pymdown-extensions>=10.14.3",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "sentry-sdk[django]>=2.20.0",
    "whitenoise>=6.9.0",
]
requires-python = ">=3.13"

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.0.1",
    "django-silk>=5.3.2",
    "django-stubs>=5.1.3",
    "mypy>=1.15.0",
]

[tool.django-stubs]
django_settings_module = "tph.settings.dev"

[tool.black]
# exclude all puzzle migration directories
extend-exclude = '^.*/migrations/'
preview = true
color = true
line-length = 88

[tool.ruff]
fix = true
line-length = 88
extend-exclude = ["**/migrations/", "puzzles/templates/make-email.py", "scripts/*"]

lint.select = [
    "E",
    "F",
    "W",
    "I",    # isort
    "DJ",   # flake8-django
    "UP",
    "Q",    # flake8-quotes
    "SIM",  # flake8-simplify
    "TC",   # flake8-type-checking,
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "PL",
    "DTZ",  # flake8-datetimez
    "ISC",  # flake8-implicit-string-concat
    "EM",   # flake8-errmsg
    "PERF",
    "FURB",
]

lint.ignore = [
    "E501",    # line length, allow literals
    "RUF012",  # incompatible with Django
    "F403",    # star import,
    "SIM102",  # nested if -- not always better
    "SIM108",  # ternary, ^
    "PLR091",  # too many whatever
    "PLR2004", # magic numbers are ok
]
